requirements.txt
django>=3.2
PyMySQL
python-dotenv
djangorestframework
PyJWT
Pillow
pytz
tzdata

